+++ client/src/app/protected/user/[id]/UserBanner.tsx
